barebone: Cast the range size XNU reports
The shared range reporter takes a u64, which is what every other
kernel's enumerator hands it; XNU's uses a usize.

+++ src/barebone/agent/src/gum_injected.rs
@@ -648,7 +648,7 @@ pub extern "C" fn _gum_process_enumerate_ranges(
         return;
     }
 
-    kernel::enumerate_ranges(&mut |base, size, protection| report(base, size, protection));
+    kernel::enumerate_ranges(&mut |base, size, protection| report(base, size as u64, protection));
 }
 
 #[cfg(feature = "linux-injected")]
